Defensive parsing of hints in oso files (#924)
This is hoped to relieve 3DSMax users of some unexplained crashes in
this section of code. (We can't reproduce, we're making our best guess.)

Related to a more extensive change under review in master (923), which
backports straightforwardly to OSL 1.9, but for this backport to OSL
1.8, we need to accommodate much older OIIO so we are more focused
here. (The other, more permanent, overhaul of this section of code
relies on string parsing that was part of OIIO 1.7, but OSL 1.8 must be
compatible with OIIO 1.6.)